twitter-python/requirements.txt
kafka-python
pandas
nltk
python-dotenv
requests
praw